Fix compilation_level=BUNDLE for closure_js_binary (#306)
closure_js_binary always advertises the sourcemap output however compiler
may not generate it (e.g. compilation_level=BUNLDLE). This patch
fixes that so the output is always generated by providing an emtpy one
if compiler didn't generate it.

+++ java/com/google/javascript/jscomp/JsCompiler.java
@@ -202,10 +202,14 @@ public final class JsCompiler implements CommandLineProgram {
       if (jsOutputFile != null) {
         Files.write(jsOutputFile, EMPTY_BYTE_ARRAY);
       }
-      if (createSourceMap != null) {
-        Files.write(createSourceMap, EMPTY_BYTE_ARRAY);
-      }
     }
+
+    // Make sure a source map is always created since Bazel expect that but JsCompiler
+    // may not emit sometimes (e.g compiler_level=BUNLDE)
+    if (createSourceMap != null && !Files.exists(createSourceMap)) {
+      Files.write(createSourceMap, EMPTY_BYTE_ARRAY);
+    }
+
     if (!failed && expectFailure) {
       System.err.println("ERROR: Expected failure but didn't fail.");
     }
